backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
google-generativeai
python-dotenv

# Development and testing
ruff
pytest
pytest-cov
httpx # For testing FastAPI endpoints
requests
